backend/requirements.txt
flask
flask-cors
deep-translator
pytesseract
opencv-python
pillow
numpy
langdetect
SpeechRecognition
gTTS
